requirements.txt
# Pydantic AI com suporte a Google Gemini
pydantic-ai[google]>=0.0.49

# Utilitários
python-dotenv>=1.0.0
rich>=13.7.0
